requirements.txt
absl-py==1.3.0
accelerate==0.18.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anytree==2.8.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.0
bert-score==0.3.13
blinker==1.4
bpemb==0.3.4
brotlipy==0.7.0
cachetools==4.2.2
cchardet==2.1.7
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.0.4
cloudpickle==2.2.1
cmake==3.26.1
colorama==0.4.6
conllu==4.5.2
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
datasets==2.4.0
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.5.1
docopt==0.6.2
editdistance==0.6.2
entrypoints==0.4
faiss-cpu==1.7.3
faiss-gpu==1.7.2
filelock==3.12.0
flair==0.11.3
flatbuffers==2.0
flit_core==3.8.0
fonttools==4.39.2
frozenlist==1.3.3
fsspec==2023.4.0
ftfy==6.1.1
future==0.18.3
gast==0.4.0
gdown==4.4.0
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.6.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.7.0
huggingface-hub==0.14.1
hyperopt==0.2.7
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
Janome==0.4.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langdetect==1.0.9
language-tool-python==2.7.1
lemminflect==0.2.3
lit==16.0.0
lru-dict==1.1.8
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
more-itertools==9.1.0
mpld3==0.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.13
networkx==3.0
nltk==3.8.1
num2words==0.5.12
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
olefile==0.46
openai==0.27.2
OpenHowNet==2.0
opt-einsum==3.3.0
overrides==3.1.0
packaging==23.0
pandas==1.5.3
pdfminer.six==20221105
Pillow==9.4.0
pinyin==0.4.0
pip==23.0.1
pkgutil_resolve_name==1.3.10
pooch==1.4.0
pptree==3.1
protobuf==3.20.3
psutil==5.9.4
py4j==0.10.9.7
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycld2==0.41
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.4.0
Pympler==1.0.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pypdf==3.8.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.3.23
requests==2.29.0
requests-oauthlib==1.3.0
responses==0.18.0
rich==13.3.4
rsa==4.7.2
scikit-learn==1.2.2
scipy==1.10.1
segtok==1.5.11
sentence-transformers==2.2.2
sentencepiece==0.1.95
setuptools==65.6.3
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
soupsieve==2.4
sqlitedict==2.1.0
streamlit==1.21.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tk==0.1.0
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchvision
tornado==6.3.1
tqdm==4.65.0
transformers==4.27.3
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
wheel==0.38.4
Wikipedia-API==0.5.8
word2number==1.1
wrapt==1.15.0
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
nltk==3.8.1
PyMuPDF==1.22.2
rank-bm25==0.2.2
python-docx==0.8.11